+++ components/Layout/Header.tsx
@@ -10,10 +10,9 @@ const Header: React.FC = () => {
     const { data: session, status } = useSession();
 
     const links: { title: string; url: string; }[] = [
-        { title: "Posts", url: "/posts" },
-        { title: "Components", url: "/compShowcase" },
-        { title: "App", url: "/app" },
-        { title: "SignUp", url: "/signUp" },
+        { title: "App", url: "./app" },
+        { title: "Archive", url: "./archive" },
+
     ]
     const image = "/keyboard.svg";
 
@@ -35,7 +34,7 @@ const Header: React.FC = () => {
     if (!session) {
         right = (
             <div>
-                <Link href="/api/auth/signin" className="hover:bg-slate-50 h-full flex text-white hover:text-violet-600 mr-4 pt-1 m-auto text-xl font-mono">
+                <Link href="./api/auth/signin" className="hover:bg-slate-50 h-full flex text-white hover:text-violet-600 mr-4 pt-1 m-auto text-xl font-mono">
                     Login
                 </Link>
             </div>
@@ -44,7 +43,7 @@ const Header: React.FC = () => {
     if (session) {
         right = (
             <div className="flex">
-                <Link href="/account" className="h-full hover:bg-slate-50 w-full flex text-white hover:text-violet-600 mr-4 pt-1 m-auto text-xl font-mono">
+                <Link href="./account" className="h-full hover:bg-slate-50 w-full flex text-white hover:text-violet-600 mr-4 pt-1 m-auto text-xl font-mono">
                     {session.user?.name}
                 </Link>
             </div>
