+++ pages/api/auth/[...nextauth].ts
@@ -7,6 +7,7 @@ import prisma from '../../../lib/prisma';
 const authHandler: NextApiHandler = (req, res) => NextAuth(req, res, options);
 export default authHandler;
 
+// This is the page that lets you log in with github.
 const options = {
   providers: [
     GitHubProvider({
